Split Page owner from Browser
