+++ lib/capybara/playwright.rb
@@ -5,6 +5,7 @@ require 'playwright'
 
 require 'capybara/playwright/browser'
 require 'capybara/playwright/browser_options'
+require 'capybara/playwright/browser_window'
 require 'capybara/playwright/dialog_event_handler'
 require 'capybara/playwright/driver'
 require 'capybara/playwright/node'
